requirements.txt
# Core dependencies
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
pydantic>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Development
pre-commit>=3.0.0
ipython>=8.0.0
